Move the DappConnectorStore to the toplevel stores
